Changed variable numbers to be int instead of long

+++ Engine/src/engine/api/dto/InstructionPeek.java
@@ -5,5 +5,5 @@ public record InstructionPeek(
         String label,
         boolean isSynthetic,
         int cycles,
-        long number
+        int number
 ){}
